+++ Kickify.Application/Features/PlayerProfiles/Commands/UpdatePlayerProfile/UpdatePlayerProfileCommand.cs
@@ -6,7 +6,7 @@ namespace Kickify.Application.Features.PlayerProfiles.Commands.UpdatePlayerProfi
     {
         public Guid ProfileId { get; set; }
         public int? CurrentElo { get; set; }
-        public decimal? TrustScore { get; set; }
+        public int? TrustScore { get; set; }
         public int? TotalMatches { get; set; }
         public int? Wins { get; set; }
         public int? Losses { get; set; }
